Add files via upload
A drunk driving detection model that uses a breathalyser to detect the level of alcohol in breath and stops the engine when the detected amount exceeds a certain value. Integrated with raspberry pi and coded in python, our model can be put to great use to prevent the number of accidents caused due to intoxicated driving.